+++ ic-agent/src/agent/mod.rs
@@ -48,7 +48,7 @@ use std::{
     collections::HashMap,
     convert::TryFrom,
     fmt,
-    future::Future,
+    future::{Future, IntoFuture},
     pin::Pin,
     sync::{Arc, Mutex, RwLock},
     task::{Context, Poll},
@@ -217,7 +217,6 @@ impl<I: Transport + ?Sized> Transport for Arc<I> {
 ///   let response = agent.update(&management_canister_id, "provisional_create_canister_with_cycles")
 ///     .with_effective_canister_id(effective_canister_id)
 ///     .with_arg(Encode!(&Argument { amount: None })?)
-///     .call_and_wait()
 ///     .await?;
 ///
 ///   let result = Decode!(response.as_slice(), CreateCanisterResult)?;
@@ -1628,7 +1627,15 @@ impl<'agent> QueryBuilder<'agent> {
     }
 }
 
-/// An in-flight canister update call. Useful primarily as a [`Future`].
+impl<'agent> IntoFuture for QueryBuilder<'agent> {
+    type IntoFuture = AgentFuture<'agent, Vec<u8>>;
+    type Output = Result<Vec<u8>, AgentError>;
+    fn into_future(self) -> Self::IntoFuture {
+        Box::pin(self.call())
+    }
+}
+
+/// An in-flight canister update call. Useful primarily as a `Future`.
 pub struct UpdateCall<'agent> {
     agent: &'agent Agent,
     response_future: AgentFuture<'agent, CallResponse<Vec<u8>>>,
@@ -1786,6 +1793,14 @@ impl<'agent> UpdateBuilder<'agent> {
             signed_update,
             request_id,
         })
+    }
+}
+
+impl<'agent> IntoFuture for UpdateBuilder<'agent> {
+    type IntoFuture = AgentFuture<'agent, Vec<u8>>;
+    type Output = Result<Vec<u8>, AgentError>;
+    fn into_future(self) -> Self::IntoFuture {
+        Box::pin(self.call_and_wait())
     }
 }