+++ ic-utils/src/interfaces/management_canister/builders.rs
@@ -4,7 +4,8 @@
 pub use super::attributes::{
     ComputeAllocation, FreezingThreshold, MemoryAllocation, ReservedCyclesLimit, WasmMemoryLimit,
 };
-use super::{ChunkHash, ManagementCanister};
+use super::{ChunkHash, LogVisibility, ManagementCanister};
+use crate::call::CallFuture;
 use crate::{
     call::AsyncCall, canister::Argument, interfaces::management_canister::MgmtMethod, Canister,
 };
@@ -20,7 +21,10 @@ use sha2::{Digest, Sha256};
 use std::{
     collections::BTreeSet,
     convert::{From, TryInto},
+    future::IntoFuture,
     pin::Pin,
+    pin::Pin,
+    str::FromStr,
     str::FromStr,
 };
 
@@ -73,6 +77,11 @@ pub struct CanisterSettings {
     ///
     /// Must be a number between 0 and 2^48^ (i.e 256TB), inclusively.
     pub wasm_memory_limit: Option<Nat>,
+
+    /// The canister log visibility of the canister.
+    ///
+    /// If unspecified and a canister is being created with these settings, defaults to `Controllers`, i.e. private by default.
+    pub log_visibility: Option<LogVisibility>,
 }
 
 /// A builder for a `create_canister` call.
@@ -86,6 +95,7 @@ pub struct CreateCanisterBuilder<'agent, 'canister: 'agent> {
     freezing_threshold: Option<Result<FreezingThreshold, AgentError>>,
     reserved_cycles_limit: Option<Result<ReservedCyclesLimit, AgentError>>,
     wasm_memory_limit: Option<Result<WasmMemoryLimit, AgentError>>,
+    log_visibility: Option<Result<LogVisibility, AgentError>>,
     is_provisional_create: bool,
     amount: Option<u128>,
     specified_id: Option<Principal>,
@@ -103,6 +113,7 @@ impl<'agent, 'canister: 'agent> CreateCanisterBuilder<'agent, 'canister> {
             freezing_threshold: None,
             reserved_cycles_limit: None,
             wasm_memory_limit: None,
+            log_visibility: None,
             is_provisional_create: false,
             amount: None,
             specified_id: None,
@@ -317,9 +328,35 @@ impl<'agent, 'canister: 'agent> CreateCanisterBuilder<'agent, 'canister> {
         }
     }
 
+    /// Pass in a log visibility setting for the canister.
+    pub fn with_log_visibility<C, E>(self, log_visibility: C) -> Self
+    where
+        E: std::fmt::Display,
+        C: TryInto<LogVisibility, Error = E>,
+    {
+        self.with_optional_log_visibility(Some(log_visibility))
+    }
+
+    /// Pass in a log visibility optional setting for the canister. If this is [None],
+    /// it will revert the log visibility to default.
+    pub fn with_optional_log_visibility<E, C>(self, log_visibility: Option<C>) -> Self
+    where
+        E: std::fmt::Display,
+        C: TryInto<LogVisibility, Error = E>,
+    {
+        Self {
+            log_visibility: log_visibility.map(|visibility| {
+                visibility
+                    .try_into()
+                    .map_err(|e| AgentError::MessageError(format!("{}", e)))
+            }),
+            ..self
+        }
+    }
+
     /// Create an [AsyncCall] implementation that, when called, will create a
     /// canister.
-    pub fn build(self) -> Result<impl 'agent + AsyncCall<(Principal,)>, AgentError> {
+    pub fn build(self) -> Result<impl 'agent + AsyncCall<Value = (Principal,)>, AgentError> {
         let controllers = match self.controllers {
             Some(Err(x)) => return Err(AgentError::MessageError(format!("{}", x))),
             Some(Ok(x)) => Some(x),
@@ -350,6 +387,11 @@ impl<'agent, 'canister: 'agent> CreateCanisterBuilder<'agent, 'canister> {
             Some(Ok(x)) => Some(Nat::from(u64::from(x))),
             None => None,
         };
+        let log_visibility = match self.log_visibility {
+            Some(Err(x)) => return Err(AgentError::MessageError(format!("{}", x))),
+            Some(Ok(x)) => Some(x),
+            None => None,
+        };
 
         #[derive(Deserialize, CandidType)]
         struct Out {
@@ -372,6 +414,7 @@ impl<'agent, 'canister: 'agent> CreateCanisterBuilder<'agent, 'canister> {
                     freezing_threshold,
                     reserved_cycles_limit,
                     wasm_memory_limit,
+                    log_visibility,
                 },
                 specified_id: self.specified_id,
             };
@@ -389,6 +432,7 @@ impl<'agent, 'canister: 'agent> CreateCanisterBuilder<'agent, 'canister> {
                     freezing_threshold,
                     reserved_cycles_limit,
                     wasm_memory_limit,
+                    log_visibility,
                 })
                 .with_effective_canister_id(self.effective_canister_id)
         };
@@ -411,15 +455,24 @@ impl<'agent, 'canister: 'agent> CreateCanisterBuilder<'agent, 'canister> {
 
 #[cfg_attr(target_family = "wasm", async_trait(?Send))]
 #[cfg_attr(not(target_family = "wasm"), async_trait)]
-impl<'agent, 'canister: 'agent> AsyncCall<(Principal,)>
-    for CreateCanisterBuilder<'agent, 'canister>
-{
-    async fn call(self) -> Result<CallResponse<(Principal,)>, AgentError> {
+impl<'agent, 'canister: 'agent> AsyncCall for CreateCanisterBuilder<'agent, 'canister> {
+    type Value = (Principal,);
+
+    async fn call(self) -> Result<CallResponse<(RequestId, _)>, AgentError> {
         self.build()?.call().await
     }
 
     async fn call_and_wait(self) -> Result<(Principal,), AgentError> {
         self.build()?.call_and_wait().await
+    }
+}
+
+impl<'agent, 'canister: 'agent> IntoFuture for CreateCanisterBuilder<'agent, 'canister> {
+    type IntoFuture = CallFuture<'agent, (Principal,)>;
+    type Output = Result<(Principal,), AgentError>;
+
+    fn into_future(self) -> Self::IntoFuture {
+        AsyncCall::call_and_wait(self)
     }
 }
 
@@ -530,7 +583,7 @@ impl<'agent, 'canister: 'agent> InstallCodeBuilder<'agent, 'canister> {
 
     /// Create an [AsyncCall] implementation that, when called, will install the
     /// canister.
-    pub fn build(self) -> Result<impl 'agent + AsyncCall<()>, AgentError> {
+    pub fn build(self) -> Result<impl 'agent + AsyncCall<Value = ()>, AgentError> {
         Ok(self
             .canister
             .update(MgmtMethod::InstallCode.as_ref())
@@ -557,13 +610,24 @@ impl<'agent, 'canister: 'agent> InstallCodeBuilder<'agent, 'canister> {
 
 #[cfg_attr(target_family = "wasm", async_trait(?Send))]
 #[cfg_attr(not(target_family = "wasm"), async_trait)]
-impl<'agent, 'canister: 'agent> AsyncCall<()> for InstallCodeBuilder<'agent, 'canister> {
+impl<'agent, 'canister: 'agent> AsyncCall for InstallCodeBuilder<'agent, 'canister> {
+    type Value = ();
+
     async fn call(self) -> Result<CallResponse<()>, AgentError> {
         self.build()?.call().await
     }
 
     async fn call_and_wait(self) -> Result<(), AgentError> {
         self.build()?.call_and_wait().await
+    }
+}
+
+impl<'agent, 'canister: 'agent> IntoFuture for InstallCodeBuilder<'agent, 'canister> {
+    type IntoFuture = CallFuture<'agent, ()>;
+    type Output = Result<(), AgentError>;
+
+    fn into_future(self) -> Self::IntoFuture {
+        AsyncCall::call_and_wait(self)
     }
 }
 
@@ -637,7 +701,7 @@ impl<'agent: 'canister, 'canister> InstallChunkedCodeBuilder<'agent, 'canister> 
     }
 
     /// Create an [`AsyncCall`] implementation that, when called, will install the canister.
-    pub fn build(self) -> Result<impl 'agent + AsyncCall<()>, AgentError> {
+    pub fn build(self) -> Result<impl 'agent + AsyncCall<Value = ()>, AgentError> {
         #[derive(CandidType)]
         struct In {
             mode: InstallMode,
@@ -686,12 +750,24 @@ impl<'agent: 'canister, 'canister> InstallChunkedCodeBuilder<'agent, 'canister> 
 
 #[cfg_attr(target_family = "wasm", async_trait(?Send))]
 #[cfg_attr(not(target_family = "wasm"), async_trait)]
-impl<'agent, 'canister: 'agent> AsyncCall<()> for InstallChunkedCodeBuilder<'agent, 'canister> {
+impl<'agent, 'canister: 'agent> AsyncCall for InstallChunkedCodeBuilder<'agent, 'canister> {
+    type Value = ();
+
     async fn call(self) -> Result<CallResponse<()>, AgentError> {
         self.call().await
     }
+
     async fn call_and_wait(self) -> Result<(), AgentError> {
         self.call_and_wait().await
+    }
+}
+
+impl<'agent, 'canister: 'agent> IntoFuture for InstallChunkedCodeBuilder<'agent, 'canister> {
+    type IntoFuture = CallFuture<'agent, ()>;
+    type Output = Result<(), AgentError>;
+
+    fn into_future(self) -> Self::IntoFuture {
+        AsyncCall::call_and_wait(self)
     }
 }
 
@@ -844,6 +920,17 @@ impl<'agent: 'canister, 'canister: 'builder, 'builder> InstallBuilder<'agent, 'c
     }
 }
 
+impl<'agent: 'canister, 'canister: 'builder, 'builder> IntoFuture
+    for InstallBuilder<'agent, 'canister, 'builder>
+{
+    type IntoFuture = CallFuture<'builder, ()>;
+    type Output = Result<(), AgentError>;
+
+    fn into_future(self) -> Self::IntoFuture {
+        Box::pin(self.call_and_wait())
+    }
+}
+
 /// A builder for an `update_settings` call.
 #[derive(Debug)]
 pub struct UpdateCanisterBuilder<'agent, 'canister: 'agent> {
@@ -855,6 +942,7 @@ pub struct UpdateCanisterBuilder<'agent, 'canister: 'agent> {
     freezing_threshold: Option<Result<FreezingThreshold, AgentError>>,
     reserved_cycles_limit: Option<Result<ReservedCyclesLimit, AgentError>>,
     wasm_memory_limit: Option<Result<WasmMemoryLimit, AgentError>>,
+    log_visibility: Option<Result<LogVisibility, AgentError>>,
 }
 
 impl<'agent, 'canister: 'agent> UpdateCanisterBuilder<'agent, 'canister> {
@@ -869,6 +957,7 @@ impl<'agent, 'canister: 'agent> UpdateCanisterBuilder<'agent, 'canister> {
             freezing_threshold: None,
             reserved_cycles_limit: None,
             wasm_memory_limit: None,
+            log_visibility: None,
         }
     }
 
@@ -1037,9 +1126,35 @@ impl<'agent, 'canister: 'agent> UpdateCanisterBuilder<'agent, 'canister> {
         }
     }
 
+    /// Pass in a log visibility setting for the canister.
+    pub fn with_log_visibility<C, E>(self, log_visibility: C) -> Self
+    where
+        E: std::fmt::Display,
+        C: TryInto<LogVisibility, Error = E>,
+    {
+        self.with_optional_log_visibility(Some(log_visibility))
+    }
+
+    /// Pass in a log visibility optional setting for the canister. If this is [None],
+    /// leaves the log visibility unchanged.
+    pub fn with_optional_log_visibility<E, C>(self, log_visibility: Option<C>) -> Self
+    where
+        E: std::fmt::Display,
+        C: TryInto<LogVisibility, Error = E>,
+    {
+        Self {
+            log_visibility: log_visibility.map(|limit| {
+                limit
+                    .try_into()
+                    .map_err(|e| AgentError::MessageError(format!("{}", e)))
+            }),
+            ..self
+        }
+    }
+
     /// Create an [AsyncCall] implementation that, when called, will update a
     /// canisters settings.
-    pub fn build(self) -> Result<impl 'agent + AsyncCall<()>, AgentError> {
+    pub fn build(self) -> Result<impl 'agent + AsyncCall<Value = ()>, AgentError> {
         #[derive(CandidType)]
         struct In {
             canister_id: Principal,
@@ -1076,6 +1191,11 @@ impl<'agent, 'canister: 'agent> UpdateCanisterBuilder<'agent, 'canister> {
             Some(Ok(x)) => Some(Nat::from(u64::from(x))),
             None => None,
         };
+        let log_visibility = match self.log_visibility {
+            Some(Err(x)) => return Err(AgentError::MessageError(format!("{}", x))),
+            Some(Ok(x)) => Some(x),
+            None => None,
+        };
 
         Ok(self
             .canister
@@ -1089,6 +1209,7 @@ impl<'agent, 'canister: 'agent> UpdateCanisterBuilder<'agent, 'canister> {
                     freezing_threshold,
                     reserved_cycles_limit,
                     wasm_memory_limit,
+                    log_visibility,
                 },
             })
             .with_effective_canister_id(self.canister_id)
@@ -1108,13 +1229,22 @@ impl<'agent, 'canister: 'agent> UpdateCanisterBuilder<'agent, 'canister> {
 
 #[cfg_attr(target_family = "wasm", async_trait(?Send))]
 #[cfg_attr(not(target_family = "wasm"), async_trait)]
-impl<'agent, 'canister: 'agent> AsyncCall<()> for UpdateCanisterBuilder<'agent, 'canister> {
+impl<'agent, 'canister: 'agent> AsyncCall for UpdateCanisterBuilder<'agent, 'canister> {
+    type Value = ();
     async fn call(self) -> Result<CallResponse<()>, AgentError> {
         self.build()?.call().await
     }
 
     async fn call_and_wait(self) -> Result<(), AgentError> {
         self.build()?.call_and_wait().await
+    }
+}
+
+impl<'agent, 'canister: 'agent> IntoFuture for UpdateCanisterBuilder<'agent, 'canister> {
+    type IntoFuture = CallFuture<'agent, ()>;
+    type Output = Result<(), AgentError>;
+    fn into_future(self) -> Self::IntoFuture {
+        AsyncCall::call_and_wait(self)
     }
 }